feature(typescript): converted: deck
