+++ src/components/deck/deck-styles.ts
@@ -1,3 +1,5 @@
+import { CSSProperties } from 'react';
+
 export function overviewFrameStyle({
   overviewScale,
   nativeSlideWidth,
@@ -6,7 +8,7 @@ export function overviewFrameStyle({
   overviewScale: number;
   nativeSlideWidth: number;
   nativeSlideHeight: number;
-}): Partial<CSSStyleDeclaration> {
+}): CSSProperties {
   return {
     margin: '1rem',
     width: `${overviewScale * nativeSlideWidth}px`,
@@ -22,7 +24,7 @@ export function overviewWrapperStyle({
   overviewScale
 }: {
   overviewScale: number;
-}): Partial<CSSStyleDeclaration> {
+}): CSSProperties {
   return {
     width: `${100 / overviewScale}%`,
     height: `${100 / overviewScale}%`,
@@ -40,7 +42,7 @@ export function printFrameStyle({
   nativeSlideWidth: number;
   nativeSlideHeight: number;
   printScale: number;
-}): Partial<CSSStyleDeclaration> {
+}): CSSProperties {
   return {
     margin: '0',
     width: `${printScale * nativeSlideWidth}px`,
@@ -57,7 +59,7 @@ export function printWrapperStyle({
   printScale
 }: {
   printScale: number;
-}): Partial<CSSStyleDeclaration> {
+}): CSSProperties {
   return {
     width: `${100 / printScale}%`,
     height: `${100 / printScale}%`,
